refactor: move testRunDatat
